+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 - Change finance handling to use `finance_subgroups` and `finance_groups` defined in the `plant_config` rather than previous `financial_groups` in the `tech_config` and `technologies_to_include_in_metrics` in `plant_config`
 - Added variable O&M to `CostModelBaseClass` and integrated into finance-related models
 - Added generic storage model, useful for battery, hydrogen, CO2, or other resource storage.
+- Added wind resource model, API baseclasses, updated examples, and documentation.
 
 ## 0.3.0 [May 2 2025]